+++ packages/synapse-interface/components/Portfolio/Portfolio.tsx
@@ -96,7 +96,7 @@ export const Portfolio = () => {
   )
 }
 
-function filterPortfolioBalancesWithBalances(
+export function filterPortfolioBalancesWithBalances(
   balancesAndAllowances: NetworkTokenBalancesAndAllowances
 ): NetworkTokenBalancesAndAllowances {
   return Object.entries(balancesAndAllowances).reduce(