+++ packages/synapse-interface/components/Portfolio/PortfolioTokenAsset.tsx
@@ -318,7 +318,7 @@ const PortfolioAssetActionButton = ({
 }: PortfolioAssetActionButtonProps) => {
   const buttonClassName = `
     flex ml-auto justify-center
-    py-1 px-6 ml-2 rounded-3xl
+    py-1 px-6 ml-2 rounded-lg
     transform-gpu transition-all duration-75
     ${isDisabled ? 'hover:cursor-default' : 'hover:cursor-pointer'}
   `