+++ packages/synapse-interface/components/ui/tailwind/Card.tsx
@@ -2,7 +2,7 @@ import { twMerge } from 'tailwind-merge'
 
 const baseClassname = `
   bg-bgBase
-  pt-3 px-6 pb-6 rounded-md
+  pt-lg px-lg pb-6 rounded-md
   `
 
 const titleBaseClassname = `