+++ packages/synapse-interface/constants/swap.ts
@@ -60,7 +60,7 @@ export const TRANSITIONS_PROPS = {
     -ml-0 md:-ml-3
     md:mt-3
     bg-bgBase
-    z-20 rounded-3xl
+    z-20 rounded-lg
   `,
 }
 
@@ -73,6 +73,6 @@ export const SETTINGS_TRANSITIONS_PROPS = {
     -ml-0 md:-ml-3
     md:-mt-3
     bg-bgBase
-    z-20 rounded-3xl
+    z-20 rounded-lg
   `,
 }