+++ packages/synapse-interface/pages/pool/poolManagement/Deposit.tsx
@@ -1,7 +1,13 @@
 import _ from 'lodash'
 
+<<<<<<< HEAD
 import { AVWETH, WETH } from '@constants/tokens/auxilliary'
 import { ETH, WETHE } from '@constants/tokens/bridgeable'
+=======
+import { WETH } from '@/constants/tokens/auxilliary'
+import { ETH, WETHE } from '@constants/tokens/bridgeable'
+import { AVWETH } from '@/constants/tokens/auxilliary'
+>>>>>>> master
 import { stringToBigInt } from '@/utils/bigint/format'
 import { DepositTokenInput } from '@components/TokenInput'
 import PriceImpactDisplay from '../components/PriceImpactDisplay'
@@ -194,7 +200,7 @@ const Deposit = ({
 
   return (
     <div className="flex-col">
-      <div className="px-2 pt-1 pb-4 bg-bgLight rounded-xl">
+      <div className="px-2 pt-1 pb-4 mb-4 rounded-md bg-bgLight">
         {pool && poolUserData.tokens && poolData ? (
           poolUserData.tokens.map((tokenObj, i) => {
             return (