+++ packages/synapse-interface/styles/transitions.ts
@@ -17,10 +17,10 @@ export interface TransitionClassesProps extends TransitionClasses {
 export const COIN_SLIDE_OVER_PROPS: TransitionClassesProps = {
   appear: true,
   unmount: true,
-  enter: 'duration-425 transition-opacity ease-out',
+  enter: 'duration-25 transition-opacity ease-out',
   enterFrom: 'transform-gpu transform opacity-0',
   enterTo: 'transform-gpu transform opacity-100',
-  leave: 'duration-350 transition-opacity ease-out',
+  leave: 'duration-25 transition-opacity ease-out',
   leaveFrom: 'transform-gpu transform opacity-100',
   leaveTo: 'transform-gpu transform opacity-0 hidden',
 }
@@ -44,7 +44,7 @@ export const TRANSITION_PROPS: TransitionClassesProps = {
     transition-all
     w-full h-full
     bg-bgBase
-    z-20 rounded-3xl
+    z-20 rounded-lg
     left-0
   `,
 }