+++ packages/synapse-interface/utils/matchSymbolWithinPool.ts
@@ -1,5 +1,5 @@
-import { WETHE, ETH } from '@constants/tokens/bridgeable'
 import { AVWETH, WETH } from '@constants/tokens/auxilliary'
+import { WETHE, ETH } from '@constants/tokens/bridgeable'
 
 /**
  * @param {Token} matchCoin the coin in the pool/ token of poolTokens